+++ .rubocop.yml
@@ -26,7 +26,7 @@ Metrics/ClassLength:
   Max: 150
 Metrics/BlockLength:
   AllowedMethods: ['describe']
-  Max: 50
+  Max: 30
 
 Style/Documentation:
   Enabled: false
